requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
openai==1.51.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
